@@ -1,31 +1,11 @@
 {
-  "name": "rs-rest-service-starter",
-  "version": "1.0.0",
-  "repository": {
-    "type": "git",
-    "url": ""
-  },
-  "type": "commonjs",
-  "description": "REST service starter for RS School",
-  "main": "src/server.js",
+  "name": "trello-competitor",
   "scripts": {
-    "start": "nodemon src/server.js",
-    "lint": "eslint --ext .js,.ts --fix src/",
+    "dev": "nodemon -w src",
+    "lint": "eslint src",
+    "start": "node --experimental-specifier-resolution node -r dotenv/config src/server.js",
     "test": "cross-env DEBUG=rs:* jest --testMatch \"<rootDir>/test/e2e/test/*.test.js\" --noStackTrace --runInBand",
     "test:auth": "cross-env DEBUG=rs:* TEST_MODE=auth jest --noStackTrace"
-  },
-  "keywords": [
-    "rs",
-    "school",
-    "rest",
-    "node",
-    "express",
-    "autotest",
-    "starter"
-  ],
-  "license": "ISC",
-  "engines": {
-    "node": ">=16.0.0 <17"
   },
   "jest": {
     "testEnvironment": "node",
@@ -33,39 +13,32 @@
       "./test/setup.js"
     ]
   },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
-  },
-  "lint-staged": {
-    "src/**/*.js": [
-      "npm run lint"
-    ]
-  },
   "dependencies": {
-    "cross-env": "7.0.3",
+    "@koa/router": "^10.1.1",
     "dotenv": "10.0.0",
-    "express": "4.17.1",
-    "swagger-ui-express": "4.1.6",
-    "uuid": "8.3.2",
-    "yamljs": "0.3.0"
+    "knex": "^0.95.14",
+    "koa": "^2.13.4",
+    "koa-bodyparser": "^4.3.0",
+    "objection": "^3.0.0",
+    "pino": "^7.5.1",
+    "pino-pretty": "^7.2.0",
+    "sqlite3": "^5.0.2",
+    "yup": "^0.32.11"
   },
   "devDependencies": {
+    "@shcherbin/eslint-config": "^0.2.0",
     "chai": "4.3.4",
-    "debug": "4.3.2",
+    "cross-env": "^7.0.3",
+    "debug": "4.3.3",
     "dirty-chai": "2.0.1",
-    "eslint": "8.2.0",
-    "eslint-config-airbnb-base": "15.0.0",
-    "eslint-config-prettier": "8.3.0",
-    "eslint-plugin-import": "2.25.3",
-    "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-prettier": "4.0.0",
-    "husky": "^4.3.8",
-    "jest": "27.3.1",
-    "lint-staged": "12.0.3",
+    "eslint": "^8.3.0",
+    "jest": "27.4.0",
     "nodemon": "2.0.15",
-    "prettier": "2.4.1",
     "supertest": "6.1.6"
-  }
+  },
+  "engines": {
+    "node": "16.x"
+  },
+  "type": "module",
+  "private": true
 }
